refactor: clear timeout logic 수정
