+++ svgr.d.ts
@@ -11,6 +11,5 @@ declare module "*.svg?url" {
 }
 
 declare module "*.png";
-declare module "*.svg";
 declare module "*.jpeg";
 declare module "*.jpg";
